+++ src/main/scala/com/caju/authorizer/repository/TransactionRepository.scala
@@ -4,7 +4,7 @@ import zio.*
 import zio.schema.*
 import zio.schema.DeriveSchema.*
 
-case class Transaction(account: String, totalAmount: Long, mcc: String, merchant: String)
+case class Transaction(account: String, totalAmount: Double, mcc: String, merchant: String)
 
 object Transaction:
 	given Schema[Transaction] = DeriveSchema.gen[Transaction]
